# Artisashop/ClientApp/src/api/models/AccountViewModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * API Artichaut
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AccountViewModel
 */
export interface AccountViewModel {
    /**
     * 
     * @type {string}
     * @memberof AccountViewModel
     */
    username?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AccountViewModel
     */
    email?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof AccountViewModel
     */
    emailConfirmed?: boolean | null;
    /**
     * 
     * @type {boolean}
     * @memberof AccountViewModel
     */
    twoFactorEnabled?: boolean | null;
}

/**
 * Check if a given object implements the AccountViewModel interface.
 */
export function instanceOfAccountViewModel(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function AccountViewModelFromJSON(json: any): AccountViewModel {
    return AccountViewModelFromJSONTyped(json, false);
}

export function AccountViewModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): AccountViewModel {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'username': !exists(json, 'username') ? undefined : json['username'],
        'email': !exists(json, 'email') ? undefined : json['email'],
        'emailConfirmed': !exists(json, 'emailConfirmed') ? undefined : json['emailConfirmed'],
        'twoFactorEnabled': !exists(json, 'twoFactorEnabled') ? undefined : json['twoFactorEnabled'],
    };
}

export function AccountViewModelToJSON(value?: AccountViewModel | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'username': value.username,
        'email': value.email,
        'emailConfirmed': value.emailConfirmed,
        'twoFactorEnabled': value.twoFactorEnabled,
    };
}